Move metriclogger to separate package
