+++ master/main.go
@@ -13,6 +13,7 @@ import (
 	"strconv"
 	"time"
 
+	"../metriclogger"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/doriandekoning/IN4392-cloud-computing-lab/graphs"
@@ -21,8 +22,6 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/levigross/grequests"
 	uuid "github.com/satori/go.uuid"
-	"github.com/shirou/gopsutil/cpu"
-	"github.com/shirou/gopsutil/mem"
 )
 
 var g graphs.Graph
@@ -57,12 +56,12 @@ func main() {
 
 	var err error
 
-	CreateMetricFolder()
-	go monitorResourceUsage()
+	metriclogger.CreateMetricFolder()
+	go metriclogger.MonitorResourceUsage()
 
 	maxWorkers, err = strconv.Atoi(os.Getenv("MAXWORKERS"))
 
-  Sess, err = session.NewSession(&aws.Config{
+	Sess, err = session.NewSession(&aws.Config{
 		Region: aws.String("us-east-1"),
 	})
 
@@ -287,32 +286,6 @@ func scaleWorkers() {
 		}
 		requestsSinceScaling = 0
 		time.Sleep(60 * time.Second)
-	}
-}
-
-func monitorResourceUsage() {
-	// TODO: This should also be implemented for the workers, which should send log this data once every 5 (?) seconds and send it once every minute (?) to the master.
-	var initial = true
-	for {
-		var err error
-
-		cpuPercent, err := cpu.Percent(0, false)
-		memstat, err := mem.VirtualMemory()
-
-		if err == nil {
-			// Initially we can log some extra system metrics.
-			if initial {
-				LogUIntMetric("master", TotalRAM, memstat.Total)
-				initial = false
-			}
-
-			LogFloatMetric("master", UsedCPUPercent, cpuPercent[0])
-			LogUIntMetric("master", AvailableRAM, memstat.Available)
-
-			//TODO: Automation experiment => Monitor in and outgoing packets?
-		}
-
-		time.Sleep(5 * time.Second)
 	}
 }
